Initial working service
This code has everything needed to create and manage tasks in a
basic way. Cancelling can't be done at the moment without changes
to Concord to deal with cancelling running tasks.

We can watch tasks that are started and report to Concord on their
failure, or run them aynchronously where the pipeline itself will
update it's status by calling Concord directly, or through another
service that can do so on its behalf.

We can also create projects, but not much else. That workflow still
needs to be reasoned about some more.

There are tests, but currently just for the observer functions.

Test coverage for the gitlab pipeline and project tasks, and general
utility functions need to be completed.